+++ packages/transport-bluetooth/src/server/methods/mod.rs
@@ -1,5 +1,9 @@
+pub mod enumerate;
 pub mod get_info;
-pub mod noop;
+pub mod start_scan;
+pub mod stop_scan;
 
+pub use self::enumerate::enumerate;
 pub use self::get_info::get_info;
-pub use self::noop::noop;
+pub use self::start_scan::start_scan;
+pub use self::stop_scan::stop_scan;