+++ packages/transport-bluetooth/src/server/mod.rs
@@ -1,8 +1,13 @@
+pub mod adapter_manager;
+pub mod connection_broadcast;
 pub mod connection_handler;
+pub mod device;
 pub mod message_handler;
 pub mod methods;
+pub mod platform;
 pub mod types;
 pub mod utils;
 
+pub use self::connection_broadcast::ConnectionBroadcast;
 pub use self::connection_handler::start_server;
 pub use self::message_handler::handle_message;