+++ packages/transport-bluetooth/src/server/types.rs
@@ -1,6 +1,10 @@
+use crate::server::{adapter_manager::AdapterError, device::TrezorDevice};
+use btleplug::api::CentralState;
+
 #[derive(serde::Serialize, Clone, Debug)]
 pub enum AbortProcess {
-    ClientDisconnected, // websocket client disconnected
+    ClientDisconnected(String), // websocket client disconnected
+    Scan,                       // stop scan
 }
 
 #[derive(serde::Serialize, Clone, Debug)]
@@ -13,6 +17,7 @@ pub enum ChannelMessage {
 #[serde(tag = "method", content = "params", rename_all = "snake_case")]
 pub enum WsRequestMethod {
     GetInfo,
+    Enumerate,
     StartScan,
     StopScan,
 }
@@ -28,10 +33,13 @@ pub struct WsRequest {
 #[serde(untagged)]
 pub enum WsResponsePayload {
     Info {
+        state: AdapterState,
         api_version: String,
         adapter_info: String,
         adapter_version: u8,
     },
+    Peripherals(Vec<TrezorDevice>),
+    Success(bool),
 }
 
 #[derive(serde::Serialize, Clone, Debug)]
@@ -46,8 +54,65 @@ pub struct WsError {
     pub error: String,
 }
 
+/// Enum converted and renamed from `btleplug::CentralState`
+#[derive(serde::Serialize, Clone, Debug)]
+#[serde(rename_all = "kebab-case")]
+pub enum AdapterState {
+    Unknown,
+    Enabled,
+    Disabled,
+    PermissionDenied,
+}
+
+impl From<CentralState> for AdapterState {
+    fn from(value: CentralState) -> Self {
+        match value {
+            CentralState::Unknown => Self::Unknown,
+            CentralState::PoweredOn => Self::Enabled,
+            CentralState::PoweredOff => Self::Disabled,
+            #[allow(unreachable_patterns)]
+            _ => Self::Unknown, // future proof in case btleplug adds other variants
+        }
+    }
+}
+
 #[derive(serde::Serialize, Clone, Debug)]
 #[serde(tag = "event", content = "payload", rename_all = "snake_case")]
-pub enum NotificationEvent {}
+pub enum NotificationEvent {
+    AdapterStateChanged {
+        state: AdapterState,
+    },
+    DeviceDiscovered {
+        id: String,
+        devices: Vec<TrezorDevice>,
+    },
+    DeviceUpdated {
+        id: String,
+        devices: Vec<TrezorDevice>,
+    },
+    DeviceConnected {
+        id: String,
+        devices: Vec<TrezorDevice>,
+    },
+    DeviceDisconnected {
+        id: String,
+        devices: Vec<TrezorDevice>,
+    },
+    DeviceRemoved {
+        id: String,
+    },
+}
 
-pub type MethodResult = Result<WsResponsePayload, Box<dyn std::error::Error>>;
+#[derive(Debug, thiserror::Error)]
+pub enum MethodError {
+    #[error("UnexpectedError: {0}")]
+    Unexpected(String),
+
+    #[error("BtleplugError: {0}")]
+    Btleplug(#[from] btleplug::Error),
+
+    #[error("AdapterError: {0}")]
+    Adapter(#[from] AdapterError),
+}
+
+pub type MethodResult = Result<WsResponsePayload, MethodError>;