+++ packages/transport-bluetooth/CHANGELOG.md
@@ -3,3 +3,4 @@
 ### 0.3.0
 
 - add websocket server
+- methods: get_info, start_scan, stop_scan, enumerate